Refactor: 우승자 구하는 메서드 테스트 WinnersTest > CarsTest 단위 테스트로 이동
